highbd_fwd_txfm_neon.c: Extract f{dct,adst,identity}16 x4 kernel
Currently we only have a "_xn" kernel with a "howmany" parameter for the
16 kernel however this does not match any of the other kernels where the
primary building-block is the simpler "_x4" kernel.

This change is primarily intended as a precursor to introducing
TRANSFORM_{COL,ROW} macros matching the lowbd code in a future commit,
since the macros can then use the "_x4" kernels universally.

Benchmarking this change on a Neoverse N2 machine with Clang 16 and GCC
12, the highbd txfm2d speed tests show a geomean 0.2% reduction in times
reported.

Change-Id: I7f4d40e4ffd6efd2752d96f38ba87eaa512d24e1