Add Neon implementation of highbd z1 predictor
Also adjust tests to add the corresponding cases for Neon by moving them
out of an AVX2-specific section.

Benchmarking on a Neoverse N2 machine with Clang 16 and GCC 12, this
change shows a geomean 60% reduction in runtime compared to the existing
C implementation, or roughly 2.5x the speed of the original
implementation.

Change-Id: Iae8f0bc01cc9fbd62a82622e1e678a2b214f2ab5