Facilitate deallocation of rusi and dgd_avg buffers
In this CL, the local buffers 'rusi' and 'rsc.dgd_avg' allocated in
av1_pick_filter_restoration() are added to `AV1LrPickStruct'. This
structure is introduced as a member in AV1_COMP to facilitate freeing
of these buffers in encoder_destroy() to prevent memory leaks in case
of early termination.

Bug: aomedia:3276
Change-Id: Ifb7e2566abbe29dd425920c36a10e1844cd3b159