rtc: Add frame-dropper support to external RTC RC
frame_drop_thresh and max_consec_drop are input
to the frame dropper.

Test coverage added for single layer.

Move the reset of gf_frame_index to where the index
is updated, this resolves an out of array bounds for
external rc with frame droppiing.

Support for the full superframe_drop mode for
spatial SVC to be added.

Change-Id: I08c2d7da20e1b260fd3413d10d548340cd681b3a