rtc: Add optimized scaler for 2:1 horiz only for Bilinear
This also fixes a bug with the aom_scaled_2d(),
for horiz or vert only scaling. The x/y_q4 was not set to 0
if the height/width is the same (no scale change).
So for 2:1 horiz scaling, it was still filtering along
the vertical direction.

Eightap case to be added.

Change-Id: I994195e578471c1e898e7fed55c71ea3bb792f8d