rtc: Allow for using golden reference in superres mode
Fix some conditions to allow superres mode in RTC
to use the golden reference.

And only release golden scaled_ref if its refreshed,
for non-svc mode. This allows to only scale the golden
reference it if is was last refreshed on previous frame.
This avoids additional reference frame scaling on every
frame, for rtc mode with superres.

Change-Id: I7c51e9e6a39b79f5889f8b6d4d81b47d20ee7a30