Fix the condition to call av1_set_vert_loop_filter_done()
In this CL, the condition to call av1_set_vert_loop_filter_done()
in enc_row_mt_worker_hook() is modified to also check the filter
levels as the same conditions are used to initialize 'lf_sync'.
This avoids the potential invalid access of 'lf_sync' in
case of an early termination.

Bug: aomedia:3276
Change-Id: Ifd5556c17e5e35faa87a5748a0650114e6eff964