Modify final refinement step in disflow
Instead of refining the flow field at pyramid level 0, skip this
step and instead refine the generated correspondences after
interpolation.

In theory, refining the generated correspondences is a bit more
accurate, as this can compensate for inaccuracies in the
interpolation step. In practice, we see only a small difference,
likely due to the number of other processing and filtering stages
between this and the eventual global motion parameter selection.

The main benefit is in encode time - pyramid level 0 has by far
the most flow vectors to refine, with one per 64 pixels of the source
image. So replacing this with one refine call per feature point, of
which there are generally many fewer than one per 64 pixels, leads to
a significant reduction in the runtime of the disflow step. This is
especially true at larger resolutions.

An alternative was also tested, where we refined both the dense
flow field and the generated correspondences, but this erased all
of the encode time savings here for essentially no further BDRATE
gain.

 Speed | BDRATE-PSNR | BDRATE-SSIM |   Enc time
-------+-------------+-------------+-------------
   1   |   -0.010%   |   -0.015%   |   -0.075%
   2   |   -0.001%   |   -0.018%   |   -0.189%
   3   |    0.000%   |   +0.035%   |   -0.466%
   4   |   -0.021%   |   -0.011%   |   -0.689%

STATS_CHANGED

Change-Id: If04c5cac6114e188462360a110e2d8378bda1a7f