rtc: Allow for single spatial layer with lower resoln
Given an initial svc setup with multiple spatial layers,
this CL would allow for user to dynamically change to
1 spatial layer and encode the lower resolutuon (base)
layer given by the scaling_factors set on the first call
of set_svc_params.

Also initialize the scaling_factors_den/num to 1 in
init_layer_context(), as this was causing a test failure
that was setting SVC without setting the scaling factors.

Change-Id: If0670574c9c562f500dc2ab3dcaaf151b80f333d