intrapred_neon.c: Use tbl in lowbd z2 predictor implementation
We can use the tbl instruction (and the 128-bit form on AArch64) to
avoid the excessive branching involved with a switch statement.

Also use widening arithmetic instructions around the shuffles rather
than a seperate widening step.

Benchmarking on a Neoverse N2 machine with Clang 16 and GCC 12, this
change gives a geomean 2.1% reduction in runtime for the z2 predictor.

Change-Id: I5ebce078e600219a23dec57cea7eba2095f47237