av1_fwd_txfm2d_neon.c: Use switch for small square problem sizes
For the smallest problem sizes we have a significant overhead from
needing to load and store the transform intermediate vectors. Avoiding
the kernel lookup and calling the kernels directly significantly
improves performance in these cases.

Benchmarking on a Neoverse N2 machine with Clang 16 and GCC 12, the
speed tests report a geomean ~22% reduction in times reported for 4x4
and a ~4.6% reduction for 8x8.

Change-Id: I5907fab09e40b1cea57c446fd7e604ae911ceae8