highbd_warp_plane_neon.c: Extract logic to header to match lowbd
In preparation for adding an SVE implementation pull the common logic
out into a new highbd_warp_plane_neon.h header. We also take this
opportunity to adjust the details of the implementation to match the
lowbd implementation. In particular to match the lowbd implementation we
now also special-case the alpha==0 and beta==0 cases.

Benchmarking this on a Neoverse N2 machine with Clang 16, the speed
tests show a geomean 4.1% reduction in times reported.

Change-Id: I24b5f894b79e2698ddf52d782be50c5662f0274e