Improve non-normative resize functions
Unlike the normative resize filter, the non-normative filter may
perform many steps, with rounding and clipping after each stage.
This introduces a lot of opportunities for rounding error to creep
into the results.

To reduce the impact of rounding errors, we make the following changes:

* Increase precision of inputs, converting from 8-, 10- or 12-bit
  inputs to an effective internal bit depth of 15 bits. This is offset
  and stored in a uint16_t, using similar logic to the offsets in the
  normative convolution functions.

* After each stage, round back to this 15-bit precision and clamp to
  the range of a uint16_t. The offset above is chosen so that we will
  almost never actually hit this clamping, it is just there to ensure
  that the intermediate results stay within 16 bits.

* Once all stages are finished, convert back to the original bit depth
  for output

This also allows most of the lowbd and highbd code to be unified into
a single code path, reducing overall code size.

STATS_CHANGED

Change-Id: I86a940bfa2009db5742eddff4ddb96c62ef0a720