Fix to aom_scaled_2d for horiz or vert-only scaling
This fixes a bug with the aom_scaled_2d(),
for horiz or vert only scaling. The x/y_q4 was not set
to 0 if the height/width is the same (no scale change).
So for 2:1 horiz scaling, it was still filtering along
the vertical direction.

This was fixed in this commit, that needed to be reverted
for a different reason:
https://aomedia-review.googlesource.com/c/aom/+/183441

Change-Id: I0f85e7e30046b6c05707eaaa132d4d77c1e07b26