Handle allocations in entropy coding stage
In this CL, the 'error' flag in 'od_ec_enc' structure which is used to
handle the allocation and reallocations of 'enc->buf' buffer in
od_ec_enc_init(), od_ec_enc_normalize() and od_ec_enc_done() functions
is checked correctly and aom_internal_error() is invoked to ensure a
graceful exit in case an error is encountered. Also modified the
relevant unit tests.

Bug: aomedia:3276
Change-Id: Ieb6b00bf2f02d29861bc9bacc38a5d3fd5586fba