Fix memory leaks from av1_choose_var_based_partitioning()
In this CL, the allocation of the local buffer 'vt' in
av1_choose_var_based_partitioning() is refactored to avoid
showing it up as a leak in case of memory allocation failure
in setup_planes(). Also modified av1_int_pro_motion_estimation()
to use thread specific 'xd->error_info' for error reporting
instead of 'cm->error'.

Bug: aomedia:3276
Change-Id: I013f619c93296b2c3d2ab881e3bf3b11397eddeb