Set initial_dimensions in encoder init
Set initial_dimensions in encoder initialization, so that
initial_dimensions are not 0, and some conditions related to
frame size and initial_dimensions comparisons in encoder config
are not triggered.

This avoids unnecessary memory allocations during encode config
calls.

Since initial_dimensions are now initialized during encoder
initialization, their values could be default (320, 240).
Therefore we need to change condition checks to compare if current
frame size is equal to initial_dimensions or not.

Change-Id: I7b2d7aa6da4e616bd5999c8089e86a445b4ac73c