rtc: Allow for using golden reference in superres mode
Fix some conditions to allow superres mode in RTC
to use the golden reference.

And only release golden scaled_ref if its refreshed,
for non-svc mode. This allows to only scale the golden
reference it if is was last refreshed on previous frame.
This avoids additional reference frame scaling on every
frame, for rtc mode with superres.

This is reland of:
https://aomedia-review.googlesource.com/c/aom/+/183463
But without the change for av1_mv_pred in nonrd_opt.h, which
was causing the test failure. This change also makes sure
nonrd_prune_ref_frame_search >= 1 is golden ref will be used.

Change-Id: Id27ac8742e3d2bd9a620d45a7841908a43a310be