intrapred_neon.c: More perf improvements for z2 predictor
The existing `shuffle` helper functions serve to shuffle the loaded data
into the right lanes and zero out the low lanes for the remainder,
however this is unnecessary since the zeroed data is already masked out
by the later BSL instructions. This means we can instead just use a
normal unmasked load instruction like we already do in the highbd
implementation.

Also replace compare into select with max: `a > b ? a : b` is just
`max(a, b)`, and avoid computing `min_base_y >> 1` in the loop body.

Benchmarking on a Neoverse N2 machine with Clang 17, this change shows a
geomean 88.06% reduction in times reported.

Change-Id: I321f2f462dbf706b9832ce07bbb45332b76c99de