Fix the error handling in av1_tf_info_alloc()
An allocation failure in av1_tf_info_alloc() is reported using
'cm->error'. This can result in a SEGV error when the function is
invoked by av1_check_initial_width() call in encoder_encode()
as 'cm->error.setjmp' is not set at this point. This CL adds early
exit in av1_tf_info_alloc() in case of an allocation failure
and the caller sites of av1_check_initial_width() is modified to
report the error using appropriate 'error_info'.

Bug: aomedia:3276
Change-Id: I8bdd570584378abe4fcebaf95731909cddb8c1e1