Fix the error handling in av1_init_mt_sync()
Currently av1_init_mt_sync() and the functions in it are using
‘cm->error’ to report an error. If an allocation from
av1_init_mt_sync() or from functions in it fails, it will result
in SEGV error as 'cm->error.setjmp' is not set at this stage.
This CL fixes this issue by adding a local setjmp target in
av1_init_mt_sync() to facilitate the error reporting. This CL also
moves av1_init_frame_mt() above av1_init_mt_sync() to fix the memory
leaks in case of an allocation failure.

Bug: aomedia:3276
Change-Id: I98704246f40d9745473e21877fce27fdc8dbb76c