Add a local setjmp target in prepare_fpmt_workers()
Allocation failures in 'av1_alloc_cdef_buffers()' when invoked from
'prepare_fpmt_workers()' is not handled correctly as 'cm->error.setjmp'
is not set. Because of this, execution will continue even after an
allocation failure, and result in SEGV error when the unallocated
buffers are accessed. This CL fixes this issue by adding a local setjmp
target in 'prepare_fpmt_workers()' to facilitate the error reporting.
Also, 'restore_workers_after_fpmt()' is called inside the setjmp target
to restore the worker states.

Bug: aomedia:3276
Change-Id: I4b4d95e72d23f131309feefb38bca4c94fd24ebc