Fix the error handling in av1_setup_sms_tree()
An allocation failure in av1_setup_sms_tree() is reported using
'cm->error'. This can result in a SEGV error when the function is
invoked by av1_init_tile_thread_data() as 'cm->error.setjmp' is
not set at this point. This CL adds early exit in
av1_setup_sms_tree() in case of an allocation failure and the
caller functions are modified to report the error using appropriate
'error_info'.

Bug: aomedia:3276
Change-Id: Idb1d87539138b662eb6f7467a2ed8a1e35914d20