Fix the AOM_SIMD_CAPS_MASK values for AVX and AVX2
The 0x100 bit in AOM_SIMD_CAPS_MASK is SSE4_2, not AVX512. So the
AOM_SIMD_CAPS_MASK value for AVX is 0x17f, not 0x07f, and the
AOM_SIMD_CAPS_MASK value for AVX2 is 0x1ff, not 0x0ff.

Also replaced the testing of SSE4_1 with SSE4_2 to enable testing
of SSE4_2 instruction set extension on CPUs that don't support AVX2.

Bug: aomedia:3472

Change-Id: I304b968dcefbd244ef866cce4caa5faf9018c8d0