rtc: Fix to external resize with use_psnr
Fix an issue with external resize with use_psnr,
which is caused by the use_rtc_tf feature.

Fix is to use crop_width/height for the aom_yv12_copy.
Also condition the realloc on change in current
resolution vs last encoded one, and remove the
check on key frame for entering the use_rtc_tf in
encoder.c

Add unittest to trigger issue.

Bug: aomedia:3531

Change-Id: I85512bd4d91cd0aff9c94b0947c72271e0053180