Fix integer overflows in update_b_sep_sym()
Fix integer overflows in update_b_sep_sym() by making `scale_threshold`
as large as possible and making `scaler` as small as possible.

Also make the same changes to update_a_sep_sym() even though they are
not needed to fix the bugs.

Update the comment that describe the range of a[*] in
update_b_sep_sym(). Add assertions to assert the ranges of a[*] and
b[*].

Bug: b:319140742
Bug: oss-fuzz:66474
Change-Id: I58934dc3a85da39f106d639ba02627776e19ac5b