cpu.cmake: Fix AArch64 compiler flag tests
The existing check_c_compiler_flag test uses a regex internally to match
against common error message strings in stderr from the compiler,
however this does not match the "invalid feature modifier" error that is
emitted by certain versions of GCC. This leads to the feature being
incorrectly enabled only to then fail to compile the library later.

To get around this, use the aom_check_source_compiles helper routine to
compile a trivial program with the flag instead since this does not
suffer the same problems and correctly identifies the features as being
not available.

Bug: aomedia:3543
Change-Id: I072281b2d3e986ee859ff8268bfad7a0fce3fd4c