aarch64_cpudetect.c: Don't read AT_HWCAPs if they are not needed
If FEAT_I8MM is not supported by the compiler then we do not test the
value from getauxval(AT_HWCAP2), leading to an unused variable warning.

To suppress this, only read the variable if we are going to use it,
which for now is only to check FEAT_I8MM. Also do similarly for AT_HWCAP
to match.

Bug: aomedia:3543
Change-Id: I855c255105cbfc448f64df317507dd9184731479
(cherry picked from commit 2036fbe2d8449c1054b4fb8bdf22c78f92655afc)