Rework downsampling pyramid setup
Instead of requesting the desired number of pyramid levels when
allocating, instead allocate all levels and fill them in on demand.
This allows much more flexibility in how pyramids are used.

Memory increase is minimal - higher pyramid levels are much smaller
than lower levels, and any unused levels are never touched and so
do not need to be assigned pages by the operating system.

Change-Id: Ib62b75d483438e0fe7b95d39989fdd925667deb7