Restore a couple of SIMD functions for x86
For normal use of libaom, we expect SSE4.1 support. Therefore in commmit
8e0b26f, we removed many SSE2 and SSSE3 functions which would never be
used on machines which support SSE4.1.

However, Valgrind on 32-bit x86 only supports up to SSSE3. This means
that the Valgrind tests became much slower, and caused timeouts.

To address this, restore the SSSE3 implementations of the lowbd forward
transform and of various CDEF functions. These are enabled only for
32-bit builds - 64-bit Valgrind does support SSE4.1, so these functions
are not required in 64-bit mode.

BUG: b/322787141
Change-Id: I9067a9dc89678a191e6e8b9148c555c24a424b1f