Disable GCC warnings in googletest headers
This is the GCC version of
https://aomedia-review.googlesource.com/c/aom/+/188601.

We may be able to fix the warnings by upgrading to a newer version of
googletest, but that requires updating some of our tests. It is easier
to just disable the -Wdeprecated-copy warning.

Use usage_exit.cc instead of usage_exit.c in C++ test targets so that we
can use the -Wno-deprecated-copy compiler option (which is valid for
C++/ObjC++ but not for C) to compile all files in the C++ test targets.

Change-Id: I2a718665c16371f9df1a5dac4ece8b96481fcda1