Add intermediate buffers for av1_compute_stats_highbd
The standard bitdepth version of av1_compute_stats makes use of
intermediate buffers dgd_avg and src_avg to precompute the subtraction
of the average. Port this optimization to the highbd version in
preparation for those buffers being used in a subsequent commit. For now
these buffers are never allocated for high bidepth.

Also remove the allocation for Neon, as the implementation doesn't make
use of them.

Change-Id: I3b646bafaf43013d508f343b318897a83c519410