Refactor ConvolveScaleTest unit tests
The filter values used in the scaling algorithm are specified in the
documentation of the inter-prediction process (chapter 7.11.3.4. [1])
which process contains the scaling algorithm. An important
characteristic of these filter values is that all of them are even
and we can leverage this information when optimising the
av1_convolve_2d_scale function. However the unit tests do not follow
this pattern, the test currently uses calculations to generate filter
values instead.

To enable us to take advantage of specific filter characteristics,
adjust the unit tests to use the predefined filter values of
av1_interp_filter_params_list defined in filter.h file.

Additionally, update convolutions parameters when testing with
different is_compound values.

[1]https://aomediacodec.github.io/av1-spec/av1-spec.pdf

Change-Id: I6ed1f7b269d139e221cbe35a573e06f0e2adaf59