rtc: Extract some speed features used in screen
Extract some features used for screen and add
speed features for them: newmv skip for flat blocks,
usage of idtx, and the min_thresh for scene detection.
This is to allow for better control for content with
mixed screen and video; also some of these may be used
for non-screen input.

No change in behavior.

Change-Id: I29307032a283d19f8899d980baa61a7d9b0748a4