dr_prediction_test.cc: Catch over-reads in zN predictors
The existing test setup does not allow us to catch code reading outside
the bounds of the input left/above arrays since the inputs are
statically allocated as part of a larger object.

To get around this, switch to creating a fresh vector of the correct
size as part of the test itself, mirroring the same approach previously
taken for the destination array.

Bug: aomedia:3571
Change-Id: If375c1d00e10ae910c878e70e801b6a49b4f329e